% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALTRE.R
\docType{package}
\name{ALTRE}
\alias{ALTRE}
\alias{ALTRE-package}
\title{Workflow: Post-alignment to altered enhancers and promoters}
\description{
This workflow takes alignment data and hotspot calls from
assays of open chromatin (ATAC-seq and Dnase-seq) and identifies
regions (enhancers and promoters) that differ based on cell and
tissue type.
}
\details{
ALTRE requires sample information CSV file, peak files (bed format),
and alignment bam files (BAM) as input

Workflow Steps: The following is the order in which the functions should be used:
(Click on function to get more detailed information)
\enumerate{

\item  \code{\link{loadPeaks}}

Takes in a sample information file (CSV), loads peak files, and outputs a
GRangesList object that holds all peaks for each sample type.

\item  \code{\link{getConsensusPeaks}}

Takes in a sample peaks list (output from loadPeaks in step 1), and outputs
consensus peaks. Consensus peaks are those present in at least N replicates.

The function outputs a list with the following slots:

\enumerate{
\item  GRangesList with consensus peaks for each sample type
\item  A statistics table summarizing how many peaks are in replicate samples
 and are called as consensus peaks
 }

\item \code{\link{combineAnnotatePeaks}}

The GRanges for all sample types are combined and annotated with
type specificity (which cell types the hotspot is present in) and
whether each region represented in the GRanges is a promoter (default: <1500bp from
a transcription start site) or an enhancer (>1500bp from a transcription
start site). Function can also merge regulatory regions that are within a specified
distance from each other.

\item \code{\link{getcounts}}

The number of reads overlapping all regions for each cell type is calculated.
The function outputs a DESeq object, summary statistics, and a density plot of the size of regulatory elements

\item \code{\link{countanalysis}}

Identify significantly altered regulatory elements (promoters or enhancers)
The function outputs a DESeq object, a summary statistics table, and a volcano plot (log2fold change vs. log2(adjusted pvalues))
}
}

