% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_compareresults.R
\name{comparePeaksAltre}
\alias{comparePeaksAltre}
\title{Comparison of methods for identifying altered regulatory regions}
\usage{
comparePeaksAltre(analysisresults, samplenames, reference,
  lfctypespecific = 1.5, lfcshared = 1.2, pvaltypespecific = 0.01,
  pvalshared = 0.05)
}
\arguments{
\item{analysisresults}{analysisresults of countanalysis.}

\item{samplenames}{vector of sample types}

\item{reference}{cell type to be considered "reference" or "reference" to which
other cell types will be compared}

\item{lfctypespecific}{log2fold change for type specific enhancers/promoters}

\item{lfcshared}{log2fold chance for shared enhancers/promoters}

\item{pvaltypespecific}{p-value for type specific enhancers/promoters}

\item{pvalshared}{p-value for shared enhancers/promoters}
}
\value{
matrix comparing the two methods of identifying altered regulatory
regions – one based on peak intensity, the other on peak presence as
determined by hotspot calling algorithms.

************* analysisresults is categaltre_peaks ***************
}
\description{
Creates a table to compare two methods of identifying altered regulatory regions
– one based on peak intensity, the other on peak presence as determined by
hotspot calling algorithms.
}
\examples{
liver_summary=resultscomparison(analysisresults_liver, reference="Stellate")
}

