% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{genomebrowvisual}
\alias{genomebrowvisual}
\title{Creates a custom track for visualization on genome browser}
\usage{
genomebrowvisual(analysisresults, lfctypespecific = 1.5, lfcshared = 1.2,
  pvaltypespecific = 0.01, pvalshared = 0.05, remove = FALSE,
  uncatagorized = "turquoise", referencespecific = "red",
  diseasespecific = "blue", shared = "purple", filteredout = "grey")
}
\arguments{
\item{analysisresults}{analysisresults of count analysis}

\item{lfctypespecific}{log2fold change for type specific enhancers/promoters}

\item{lfcshared}{log2fold chance for shared enhancers/promoters}

\item{pvaltypespecific}{p-value for type specific enhancers/promoters}

\item{pvalshared}{p-value for shared enhancers/promoters}

\item{remove}{removes filtered out regions}

\item{uncatagorized}{color of track for uncategorized regions}

\item{referencespecific}{color of track for reference specific regions}

\item{diseasespecific}{color of track for disease specific regions}

\item{shared}{color of track for shared regions}

\item{filteredout}{color of track for regions that are filtered out}
}
\description{
Creates a colors coded HOTSPOT (bed) file for visualization in UCSC genome
browser: red indicates increased log2fold change/low p-value, blue indicates
decreased lod2fold change/low p-value, and purple indicates regions
with little to no change and insignificant p-values. Based on the log2fold
change and p-value inputs, there is a possibility that some regions will not
fulfill any of the supplied criteria ("in-between" shared and type-specific) –
they are colored grey.
}
\examples{
\dontrun{
genomebrowvisual(categaltre_peaks)
}

}

