% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{grangestodataframe}
\alias{grangestodataframe}
\title{Internal function: converts granges to dataframe}
\usage{
grangestodataframe(grange)
}
\arguments{
\item{grange}{Grange}
}
\value{
dataframe

Dataframe1 <- grangestodataframe(Grange1)
}
\description{
Takes a Grange input and converts it into a dataframe
Not for use by package user
}
