% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingfuntions.R
\name{loadBamFiles}
\alias{loadBamFiles}
\title{Read in BAM files (internal function)}
\usage{
loadBamFiles(csvfile)
}
\arguments{
\item{csvfile}{csvfile}
}
\description{
Read in BAM files (internal function)
}

