% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_getfiles.R
\name{loadBedFiles}
\alias{loadBedFiles}
\title{Read in BED Files (internal function)}
\usage{
loadBedFiles(csvfile)
}
\arguments{
\item{csvfile}{csvfile}
}
\description{
Read in BED Files (internal function)
}

