% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_loadingfuntions.R
\name{loadPeaks}
\alias{loadPeaks}
\title{Read in open chromatin data peak files (bed format, 0-based)}
\usage{
loadPeaks(csvfile)
}
\arguments{
\item{csvfile}{name of CSV file, with complete path, that contains
the following 4 columns for each sample
 1) complete filepath;
 2) name of bamfiles;
 3) name of peak files;
 4) name or type of sample;
 5) name of replicate}
}
\value{
a GRangesList object comprising one GRanges object for each sample.
}
\description{
Read in open chromatin data peak files (bed format, 0-based)
}
\examples{
dir <- system.file('extdata', package='ALTRE', mustWork=TRUE)
csvfile <- file.path(dir, 'lung.csv')
samplePeaks <- loadPeaks(csvfile)

}

