% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{mergeclosepeaks}
\alias{mergeclosepeaks}
\title{Internal function: merges peaks within user input distance of each other}
\usage{
mergeclosepeaks(peaklist, grange, mergedist, TSS, distancefromTSS)
}
\arguments{
\item{peaklist}{peaklist}

\item{grange}{data frame of GRanges}

\item{mergedist}{numeric}

\item{TSS}{TSS}

\item{distancefromTSS}{distancefromTSS}
}
\value{
GRanges object
}
\description{
Takes a data frame of GRanges and distance as input
Not for use by package user
}
