% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{list of plots}

\item{plotlist}{plotlist}

\item{file}{file}

\item{cols}{number of columns in layout}

\item{layout}{matrix specifying layout, if present cols is ignored}
}
\description{
Plots multiple ggplot objects in one window.
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
This function was not written by the authours of this package. Link:
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}

