% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_getpathways.R
\name{pathenrich}
\alias{pathenrich}
\title{Enrichment analysis to identify putative pathways of interest for further
investigation}
\usage{
pathenrich(analysisresults, ontoltype = "MF", enrichpvalfilt = 0.01,
  lfctypespecific = 1.5, lfcshared = 1.2, pvaltypespecific = 0.01,
  pvalshared = 0.05, genes = 20, offspring = 300,
  regionsubset = "TSS-distal")
}
\arguments{
\item{analysisresults}{Results from analysis of counts, categaltre_peaks.}

\item{ontoltype}{One of three categories: 'MF' (molecular function),
'CC' (cellular component), 'BP' (biological process).}

\item{enrichpvalfilt}{Adjusted pval for enrichment to filter on
(adjusted for multiple testing).}

\item{lfctypespecific}{Log2fold change (of chromatin accessibility)
for type specific TSS-proximal and TSS-distal regions.}

\item{lfcshared}{Log2fold change (of chromatin accessibility) for
shared TSS-proximal and TSS-distal regions.}

\item{pvaltypespecific}{P-value (of chromatin accessibility) for
type specific TSS-proximal and TSS-distal regions.}

\item{pvalshared}{P-value (of chromatin accessibility) for
shared TSS-proximal and TSS-distal regions.}

\item{genes}{Minimum number of genes allowable in a pathway.}

\item{offspring}{Maximum number of offspring allowable in a pathway.}

\item{regionsubset}{'TSS-proximal' or 'TSS-distal'. Default is "TSS-distal".}
}
\value{
dataframe identifying p-values for enriched pathways --
pathways also annotated with additional information
}
\description{
Determine which pathways are overrepresented in
altered TSS-proximal and TSS-distal regions. Pathways are determined by linking
the TSS-proximal/distal regions to the nearest gene, and then linking genes to pathways
using the gene ontology database. The 'gene' argument limits how few genes
a pathway can contain, while the 'offspring' argument limits how many
offspring a pathway can contain. Pathways with low gene counts are less
reliable (often false positives), while pathways with many offspring are
vague and unlikely to be of much use -- the enrichment of their more precise
 offspring is the more interesting question.
}
\details{
************CHANGE ANALYSISRESULTS TO ALTREPEAKS!!!!******************
****************CHANGE REGIONSUBSET TO REGIONTYPE!!!!*****************
}
\examples{
\dontrun{
csvfile <- file.path(dir="yourfilepath", 'sampleinfo.csv')
sampleinfo <- loadCSVFile(csvfile)
samplePeaks <- loadBedFiles(sampleinfo)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
TSSannot <- getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consPeaks,
                                          TSS = TSSannot,
                                          merge = TRUE,
                                          regionspecific = TRUE,
                                          distancefromTSSdist = 1500,
                                          distancefromTSSprox = 1000 )
#Need to run getcounts on all chromosomes
counts_consPeaks <- getCounts(annotpeaks = consPeaksAnnotated,
                             sampleinfo = sampleinfo,
                             reference = 'SAEC')
altre_peaks <- countanalysis(counts=counts_consPeaks,
                             pval=0.01,
                             lfcvalue=1)
MFenrich <- pathenrich(analysisresults = altre_peaks,
                       ontoltype = 'MF',
                       enrichpvalfilt = 0.01)
BPenrich <- pathenrich(analysisresults=altre_peaks,
                       ontoltype='BP',
                       enrichpvalfilt=0.01)
}
}

