% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotallvenn}
\alias{plotallvenn}
\title{Plots venn diagrams for comparison of two methods of identifying altered
regulatory regions Makes venn diagrams for TSS-proximal, TSS-distal, and combined
for both intensity-based peaks and for peaks identified by hotspot calling
algorithms.  There is no return value. Six venn diagrams will be plotted}
\usage{
plotallvenn(analysisresultsmatrix)
}
\arguments{
\item{analysisresultsmatrix}{analysisresults of countanalysis function
place into a a analysisresults matrix by the analyzeanalysisresults function}
}
\description{
Plots venn diagrams for comparison of two methods of identifying altered
regulatory regions Makes venn diagrams for TSS-proximal, TSS-distal, and combined
for both intensity-based peaks and for peaks identified by hotspot calling
algorithms.  There is no return value. Six venn diagrams will be plotted
}
\examples{
\dontrun{
csvfile <- file.path(dir="yourfilepath", 'sampleinfo.csv')
sampleinfo <- loadCSVFile(csvfile)
samplePeaks <- loadBedFiles(sampleinfo)
consPeaks <- getConsensusPeaks(samplepeaks = samplePeaks, minreps = 2)
plotConsensusPeaks(samplepeaks = consPeaks)
TSSannot <- getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consPeaks,
                                           TSS = TSSannot,
                                           merge = TRUE,
                                           regionspecific = TRUE,
                                           distancefromTSSdist = 1500,
                                           distancefromTSSprox = 1000 )
counts_consPeaks <- getCounts(annotpeaks = consPeaksAnnotated,
                              sampleinfo = sampleinfo,
                              reference = 'SAEC')
altre_peaks <- countanalysis(counts=counts_consPeaks,
                             pval=0.01,
                             lfcvalue=1)
categaltre_peaks=categAltrePeaks(altre_peaks,
     lfctypespecific=1.5,
     lfcshared=1.2,
     pvaltypespecific=0.01,
     pvalshared=0.05)
analysisresults <- resultsComparison(altre_peaks, reference= "SAEC")
plotallvenn(analysisresults)
}
}

