% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotConsensusPeaks}
\alias{plotConsensusPeaks}
\title{Given the output from getConsensusPeaks, generate a barplot
of countstatistics}
\usage{
plotConsensusPeaks(samplepeaks, palette = "Set1")
}
\arguments{
\item{samplepeaks}{output generated from getConsensusPeaks}

\item{palette}{RColorBrewer palette to change graph colors}
}
\value{
a highcharter object
}
\description{
Given the output from getConsensusPeaks, generate a barplot
of countstatistics
}
\examples{
\dontrun{
csvfile <- loadCSVFile("DNAseEncodeExample.csv")
samplePeaks <- loadBedFiles(csvfile)
consensusPeaks <- getConsensusPeaks(samplepeaks = samplePeaks,
minreps = 2)
plotConsensusPeaks(samplepeaks = consensusPeaks)}
}

