% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotCountAnalysisTemp}
\alias{plotCountAnalysisTemp}
\title{Create a volcano plot from the output of categAltrePeaks}
\usage{
plotCountAnalysisTemp(altrepeakscateg = NULL, viewer = TRUE,
  palette = NULL)
}
\arguments{
\item{altrepeakscateg}{output generated from countanalysis() then
categAltrePeaks()}

\item{viewer}{whether the plot should be displayed in the RStudio viewer or
in Shiny/Knittr}

\item{palette}{RColorBrewer palette to change graph colors}
}
\value{
a highcharter object
}
\description{
Create a volcano plot from the output of categAltrePeaks
}

