% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotDistCountAnalysis}
\alias{plotDistCountAnalysis}
\title{Creates a boxplot to see the distribution of read counts in type-specific and
shared TSS-proximal and TSS-distal regions.}
\usage{
plotDistCountAnalysis(analysisresults, counts, palette = NULL,
  xlabelsize = "15px", ylabel = "log2(FPKM)", ylabelsize = "15px",
  maintitle = "Distribution of Normalized Counts (peaks types determine by intensity)",
  maintitlesize = "20px", xlabel = NULL)
}
\arguments{
\item{analysisresults}{output generated from countanalysis() then
categAltrePeaks()}

\item{counts}{output generated from getCounts()}

\item{palette}{RColorBrewer palette to change graph colors}

\item{xlabelsize}{size of xlabel (default, 15px)}

\item{ylabel}{label for y-axis (default, "Observations")}

\item{ylabelsize}{size of ylabel (default, 15px)}

\item{maintitle}{main title (default, "Distribution of Normalized Counts")}

\item{maintitlesize}{main title size (default, 20px)}

\item{xlabel}{label for x-axis (default, sample names)}
}
\value{
a highcharter object
}
\description{
Takes the rlog transformation of the RRKM (Reads Per Kilobase of transcript
per Million) of the read counts of type-specific and shared regulatory
regions and plots the distribution of those read counts in all sample types
analyzed in the workflow.
}
\examples{
\dontrun{
dir <- system.file('extdata', package='ALTRE', mustWork=TRUE)
csvfile <- file.path(dir, 'lung.csv')
sampleinfo <- loadCSVFile(csvfile)
samplePeaks <- loadBedFiles(sampleinfo)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
plotConsensusPeaks(samplepeaks=consPeaks)
TSSannot<- getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consPeaks,
                                          TSS = TSSannot,
                                          merge = TRUE,
                                          regionspecific = TRUE,
                                          distancefromTSSdist = 1500,
                                          distancefromTSSprox = 1000)
counts_consPeaks <- getCounts(annotpeaks = consPeaksAnnotated,
                              sampleinfo = sampleinfo,
                              reference = 'SAEC',
                              chrom = 'chr21')
altre_peaks <- countanalysis(counts = counts_consPeaks,
                             pval = 0.01,
                             lfcvalue = 1)
categaltre_peaks <- categAltrePeaks(altre_peaks,
                                    lfctypespecific = 1.5,
                                    lfcshared = 1.2,
                                    pvaltypespecific = 0.01,
                                    pvalshared = 0.05)
plotDistCountAnalysis(categaltre_peaks, counts_consPeaks)
}
}

