% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotGREATenrich}
\alias{plotGREATenrich}
\title{Given the output from processPathways(), creates a heatmap from
the ouput of the GREAT enrichment analysis. Presence or absence of
the pathway in enrichment of both type-specific (increased or decreased
log2fold change, low p-value) and shared (no change, higher p-value)
regulatory regions is plotted.}
\usage{
plotGREATenrich(input, maintitle = "GREAT Enrichment Analysis",
  pathwaycateg = NULL, test = "Binom", numshow = 10,
  maintitlesize = "20px", ylabelsize = "10px", xlabelsize = "10px",
  xlabel = NULL, subtitle = "(color corresponds to p-value)",
  subtitlesize = "13px")
}
\arguments{
\item{input}{results from GREAT enrichment analysis}

\item{maintitle}{main title (default, "GREAT Enrichment Analysis")}

\item{pathwaycateg}{ontology, to see available ontologies in your input results (e.g. named
GREATpathways, type getOntologies(GREATpathways)}

\item{test}{character, "Binom" uses binomial test restuls, "Hyper" uses
hypergeometric test results.  Default is "Binom"}

\item{numshow}{number of top pathways (ranked according to p-value) of each type
(expt, reference, shared) to show in the plot (default=10)}

\item{maintitlesize}{main title size (default, 20px)}

\item{ylabelsize}{size of ylabel (default, 10px)}

\item{xlabelsize}{size of xlabel (default, 10px)}

\item{xlabel}{label for x-axis (default, Experiment-specific, shared, Reference-specific )}

\item{subtitle}{subtitle (default, "color corresponds to p-value")}

\item{subtitlesize}{subitle size (default 15px)}
}
\value{
heatmap
}
\description{
Given the output from processPathways(), creates a heatmap from
the ouput of the GREAT enrichment analysis. Presence or absence of
the pathway in enrichment of both type-specific (increased or decreased
log2fold change, low p-value) and shared (no change, higher p-value)
regulatory regions is plotted.
}
\examples{
\dontrun{
csvfile <- loadCSVFile("DNAseEncodeExample.csv")
samplePeaks <- loadBedFiles(csvfile)
consensusPeaks <- getConsensusPeaks(samplepeaks = samplePeaks, minreps = 2)
TSSannot <- getTSS()
consensusPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consensusPeaks,
                                          TSS = TSSannot,
                                          merge = TRUE,
                                          regionspecific = TRUE,
                                          distancefromTSSdist = 1500,
                                          distancefromTSSprox = 1000)
consensusPeaksCounts <- getCounts(annotpeaks = consensusPeaksAnnotated,
                              reference = 'SAEC',
                              sampleinfo = csvfile,
                              chrom = 'chr21')
alteredPeaks <- countanalysis(counts=consensusPeaksCounts,
                             pval=0.01,
                             lfcvalue=1)
alteredPeaksCategorized <- categAltrePeaks(analysisresults = alteredPeaks,
                             lfctypespecific = 1.5,
                             lfcshared = 1.2,
                             pvaltypespecific = 0.01,
                             pvalshared = 0.05)
callPaths <- runGREAT(peaks = alteredPeaksCategorized)
pathResults <- processPathways(callPaths, pathway_category = "GO",
enrichcutoff = 2, adjpvalcutoff = 0.05)
plotGREATenrich(pathResults, maintitle = "GREAT Enrichment Analysis",
pathwaycateg = "GO_Molecular_Function")
}
}

