% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotvenn}
\alias{plotvenn}
\title{Plots a venn digram to compare sample types Plots the number of type-specific
and shared regulatory regions in a venn diagram Type of regulatroy region
(enhancer, promoter, or both) and type  of peak comparison (intensity or peak)
must be specified.}
\usage{
plotvenn(analysisresultsmatrix, region = "both", method = "intensity",
  color = "redorange")
}
\arguments{
\item{analysisresultsmatrix}{analysisresults of Intensity analysis place into
analysisresults matrix by the analyzeanalysisresults function}

\item{region}{pick a region, regions can be 'enhancer', 'promoter', or 'both'
INCLUDE quotes}

\item{method}{pick a method, methods can be 'intensity' or 'peak'
include quotes}

\item{color}{include the colors you want in your venn diagram}
}
\value{
venn diagram
}
\description{
Plots a venn digram to compare sample types Plots the number of type-specific
and shared regulatory regions in a venn diagram Type of regulatroy region
(enhancer, promoter, or both) and type  of peak comparison (intensity or peak)
must be specified.
}
\examples{
\dontrun{
dir <- system.file('extdata', package='ALTRE', mustWork=TRUE)
csvfile <- file.path(dir, 'lung.csv')
sampleinfo <- loadCSVFile(csvfile)
samplePeaks <- loadBedFiles(sampleinfo)
consPeaks <- getConsensusPeaks(samplepeaks = samplePeaks, minreps = 2)
plotConsensusPeaks(samplepeaks = consPeaks)
TSSannot <- getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consPeaks,
                                           TSS = TSSannot,
                                           merge = TRUE,
                                           regionspecific = TRUE,
                                           mergedistenh = 1500,
                                           mergedistprom = 1000)
counts_consPeaks <- getcounts(annotpeaks = consPeaksAnnotated,
                              csvfile = csvfile,
                              reference = 'SAEC')
altre_peaks <- countanalysis(counts=counts_consPeaks,
                             pval=0.01,
                             lfcvalue=1)
analysisresults <- resultsComparison(altre_peaks, reference= "A549")
plot1 <- plotvenn(analysisresults,
                  region='enhancer',
                  method='intensity')
}
}

