% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8a_runGREAT.R
\name{processPathways}
\alias{processPathways}
\title{Enrichment analysis using GREAT package
to identify putative pathways of interest for further
investigation}
\usage{
processPathways(GREATpath, pathway_category = "GO", adjpvalcutoff = 0.05)
}
\arguments{
\item{GREATpath}{output of runGREAT()}

\item{pathway_category}{character, "GO", "Pathway Data", "Regulatory Motifs",
"Phenotype Data and Human Disease", "Gene Expression", "Gene Families"
(default is "GO")}

\item{adjpvalcutoff:}{numeric, adjusted p-value cutoff to determine enriched pathways,
default is 0.05}
}
\value{
list of dataframes for enriched pathways - each dataframe in the list
represents one pathway type (e.g. "GO MOlecular Function")
}
\description{
Enrichment analysis using GREAT package
to identify putative pathways of interest for further
investigation
}

