% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundose.R
\name{rundose}
\alias{rundose}
\title{Internal function: runs enrichment analysis}
\usage{
rundose(set, background, log2FoldChange, pvalfilt, ontoltype, genes, offspring)
}
\arguments{
\item{set}{set}

\item{background}{background}

\item{log2FoldChange}{log2FoldChange}

\item{pvalfilt}{pvalfilt}

\item{ontoltype}{ontoltype}

\item{genes}{genes}

\item{offspring}{offspring}
}
\value{
dataframe
}
\description{
Not for use by package user
}

