% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8a_runGREAT.R
\name{runGREAT}
\alias{runGREAT}
\title{Enrichment analysis using GREAT package
to identify putative pathways of interest for further
investigation}
\usage{
runGREAT(peaks, species = "hg19", rule = "basalPlusExt", adv_upstream = 5,
  adv_downstream = 1, adv_span = 1000, adv_twoDistance = 1000,
  adv_oneDistance = 1000, pathway_category = "GO")
}
\arguments{
\item{peaks}{list, output of categAltrePeaks() function}

\item{species}{default hg19}

\item{rule}{character, "basalPlusExt", "twoClosest", "oneClosest" rule that associates
genomic regions to genes (default is "basalPlusExt").
See https://bioconductor.org/packages/release/bioc/html/chipenrich.html for more detail.}

\item{peaktype}{character, "Experiment Specific", "Reference Specific",
"Ambiguous", "Shared", or "All" (All is default)}

\item{adv_upstream:}{kb, extension to upstream (if rule is basalPlusExt), default 5}

\item{adv_downstream:}{kb, extension to downstream (if rule is basalPlusExt), default 1.0}

\item{adv_span:}{kb, max extension (if rule is basalPlusExt), default 1000.0}

\item{adv_twoDistance:}{kb, max extension (if rule is twoClosest), default 1000.0}

\item{adv_oneDistance:}{kb, max extension (if rule is oneClosest), default 1000.0}

\item{pathway_category:}{character, "GO", "Pathway Data", "Regulatory Motifs",
"Phenotype Data and Human Disease", "Gene Expression", "Gene Families"
(default is "GO")}
}
\value{
ways --
pathways also annotated with additional information
}
\description{
Enrichment analysis using GREAT package
to identify putative pathways of interest for further
investigation
}
\examples{
\dontrun{
runGREAT(peaks=categaltre_peaks)
}
}

