% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{statscombineannotate}
\alias{statscombineannotate}
\title{Internal function: creates a stat matrix for combined and annotated peaks}
\usage{
statscombineannotate(tableofinfo, input, row1, row2)
}
\arguments{
\item{tableofinfo}{tableinfo}

\item{input}{input}

\item{row1}{row1}

\item{row2}{row2}
}
\value{
matrix
}
\description{
Takes a data frame of GRanges
Not for use by package user
}

