% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writerfunctions.R
\name{writeConsensusRE}
\alias{writeConsensusRE}
\title{Writes annotated genomic regions}
\usage{
writeConsensusRE(countsPeaks, con)
}
\arguments{
\item{countsPeaks}{output generated from getCounts()}

\item{con}{connection filepath}
}
\description{
Creates a CSV file with genomic regions annoted as TSS-proximal and TSS-distal regions.
categorized by cell type
}
\examples{
\dontrun{
csvfile <- file.path(dir="yourfilepath", 'sampleinfo.csv')
sampleinfo <- loadCSVFile(csvfile)
samplePeaks <- loadBedFiles(sampleinfo)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
TSSannot <- getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consPeaks,
                                          TSS = TSSannot,
                                          merge = TRUE,
                                          regionspecific = TRUE,
                                          distancefromTSSdist = 1500,
                                          distancefromTSSprox = 1000 )
<<<<<<< HEAD:man/writeAnnotatedRegions.Rd
#Need to run getcounts on all chromosomes
countsPeaks <- getcounts(annotpeaks = consPeaksAnnotated,
                             csvfile = csvfile,
                             reference = 'SAEC')
=======
>>>>>>> webgraphs:man/writeConsensusRE.Rd
con <- "annotatedRegions.csv"
writeConsensusRE(consPeaksAnnotated, con)
}
}

