% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writerfunctions.R
\name{writeGREATpath}
\alias{writeGREATpath}
\title{Writes a data generated by pathenrich}
\usage{
writeGREATpath(pathenrichOut, con)
}
\arguments{
\item{pathenrichOut}{output generated from pathenrich()}

\item{con}{connection filepath}
}
\description{
Creates  three CSV files containing the results of the pathway enrichment analysis for
experiment-specific, reference-specific, and shared REs.
}
\examples{
\dontrun{
csvfile <- file.path(dir="yourfilepath", 'sampleinfo.csv')
sampleinfo <- loadCSVFile(csvfile)
samplePeaks <- loadBedFiles(sampleinfo)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
TSSannot <- getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consPeaks,
                                          TSS = TSSannot,
                                          merge = TRUE,
                                          regionspecific = TRUE,
                                          distancefromTSSdist = 1500,
                                          distancefromTSSprox = 1000)
#Need to run getcounts on all chromosomes
counts_consPeaks <- getcounts(annotpeaks = consPeaksAnnotated,
                             csvfile = csvfile,
                             reference = 'SAEC')
                             #' altre_peaks <- countanalysis(counts = counts_consPeaks,
                             pval = 0.01,
                             lfcvalue = 1)
categaltre_peaks <- categAltrePeaks(altre_peaks,
                                    lfctypespecific = 1.5,
                                    lfcshared = 1.2,
                                    pvaltypespecific = 0.01,
                                    pvalshared = 0.05)
callGREAT <- runGREAT(peaks=categaltre_peaks)
pathGREAT <- processPathways(callGREAT)
con <- "GREAToutput.zip"
writeGREATpath(pathenrichOut = pathGREAT, con = con)
}

}

