% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categaltrepeaks.R
\name{categAltrePeaks}
\alias{categAltrePeaks}
\title{Categorize altered peaks as experiment-specific, reference-specific,
 or shared}
\usage{
categAltrePeaks(analysisresults, lfctypespecific = 1.5, lfcshared = 1.2,
  pvaltypespecific = 0.01, pvalshared = 0.05)
}
\arguments{
\item{analysisresults}{output of countanalysis()}

\item{lfctypespecific}{log2fold change cutoff (of chromatin accessibility)
for type specific enhancers/promoters}

\item{lfcshared}{log2fold change cutoff (of chromatin accessibility)
for shared enhancers/promoters}

\item{pvaltypespecific}{p-value cutoff (of chromatin accessibility)
for type specific enhancers/promoters}

\item{pvalshared}{p-value cutoff (of chromatin accessibility)
for shared enhancers/promoters}
}
\value{
list 1) results of countanalysis,
2) Frequences of sample type specific, shared,
 and ambiguous regulatory elements
}
\description{
Categorize altered peaks as experiment-specific, reference-specific,
 or shared
}
\examples{
\dontrun{
dir <- system.file('extdata', package='ALTRE', mustWork=TRUE)
csvfile <- file.path(dir, 'lung.csv')
sampleinfo <- loadCSVFile(csvfile)
samplePeaks <- loadBEDFiles(csvfile)
counts_consPeaks <- getcounts(annotpeaks = consPeaksAnnotated,
                              sampleinfo = sampleinfo,
                              reference = 'SAEC')
altre_peaks <- countanalysis(counts = counts_consPeaks,
                             pval = 0.01,
                             lfcvalue = 1)
categaltre_peaks <- categAltrePeaks(
  altre_peaks,
  lfctypespecific = 1.5,
  lfcshared = 1.2,
  pvaltypespecific = 0.01,
  pvalshared = 0.05
)
}
}

