% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_getcategorized.R
\name{categAltrePeaks}
\alias{categAltrePeaks}
\title{Categorize altered peaks as experiment-specific, reference-specific,
 or shared}
\usage{
categAltrePeaks(analysisresults, lfctypespecific = 1.5, lfcshared = 1.2,
  pvaltypespecific = 0.01, pvalshared = 0.05)
}
\arguments{
\item{analysisresults}{output of countanalysis()}

\item{lfctypespecific}{log2fold change cutoff (of chromatin accessibility)
for type specific TSS-distals/TSS-proximals}

\item{lfcshared}{log2fold change cutoff (of chromatin accessibility)
for shared TSS-distals/TSS-proximals}

\item{pvaltypespecific}{p-value cutoff (of chromatin accessibility)
for type specific TSS-distals/TSS-proximals}

\item{pvalshared}{p-value cutoff (of chromatin accessibility)
for shared TSS-distals/TSS-proximals}
}
\value{
list 1) results of countanalysis,
2) Frequences of sample type specific, shared,
 and ambiguous regulatory elements
}
\description{
Categorize altered peaks as experiment-specific, reference-specific,
 or shared
}
\examples{
\dontrun{
csvfile <- loadCSVFile("DNaseEncodeExample.csv")
samplePeaks <- loadBedFiles(csvfile)
consensusPeaks <- getConsensusPeaks(samplepeaks = samplePeaks, minreps = 2)
TSSannot <- getTSS()
consensusPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consensusPeaks,
   TSS = TSSannot,
   merge = TRUE,
   regionspecific = TRUE,
   distancefromTSSdist = 1500,
   distancefromTSSprox = 1000)
consensusPeaksCounts <- getCounts(annotpeaks = consensusPeaksAnnotated,
   sampleinfo = csvfile,
   reference = 'SAEC',
   chrom = 'chr21')
alteredPeaks <- countanalysis(counts = consensusPeaksCounts,
   pval = 0.01,
   lfcvalue = 1)
alteredPeaksCategorized <- categAltrePeaks(analysisresults = alteredPeaks,
   lfctypespecific = 1.5,
   lfcshared = 1.2,
   pvaltypespecific = 0.01,
   pvalshared = 0.05)
}
}
