% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_getcomparison.R
\name{comparePeaksAltre}
\alias{comparePeaksAltre}
\title{Comparison of methods for identifying altered regulatory regions}
\usage{
comparePeaksAltre(analysisresults, lfctypespecific = 1.5, lfcshared = 1.2,
  pvaltypespecific = 0.01, pvalshared = 0.05)
}
\arguments{
\item{analysisresults}{analysisresults of countanalysis.}

\item{lfctypespecific}{log2fold change for type specific TSS-dists/TSS-proxs}

\item{lfcshared}{log2fold chance for shared TSS-dists/TSS-proxs}

\item{pvaltypespecific}{p-value for type specific TSS-dists/TSS-proxs}

\item{pvalshared}{p-value for shared TSS-dists/TSS-proxs}
}
\value{
matrix comparing the two methods of identifying altered regulatory
regions, one based on peak intensity, the other on peak presence as
determined by hotspot calling algorithms.
}
\description{
Creates a table to compare two methods of identifying altered regulatory
regions, one based on peak intensity, the other on peak presence as
 determined by hotspot calling algorithms.
}
\examples{
\dontrun{
csvfile <- loadCSVFile("DNAseEncodeExample.csv")
samplePeaks <- loadBedFiles(csvfile)
consensusPeaks <- getConsensusPeaks(samplepeaks = samplePeaks, minreps = 2)
TSSannot <- getTSS()
consensusPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consensusPeaks,
   TSS = TSSannot,
   merge = TRUE,
   regionspecific = TRUE,
   distancefromTSSdist = 1500,
   distancefromTSSprox = 1000)
consensusPeaksCounts <- getCounts(annotpeaks = consensusPeaksAnnotated,
   sampleinfo = csvfile,
   reference = 'SAEC',
   chrom = 'chr21')
alteredPeaks <- countanalysis(counts = consensusPeaksCounts,
   pval = 0.01,
   lfcvalue = 1)
alteredPeaksCategorized <- categAltrePeaks(alteredPeaks,
   lfctypespecific = 1.5,
   lfcshared = 1.2,
   pvaltypespecific = 0.01,
   pvalshared = 0.05)
comparePeaksAnalysisResults <- comparePeaksAltre(alteredPeaksCategorized)
}
}

