% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_getcomparison.R
\name{comparePeaksAltre}
\alias{comparePeaksAltre}
\title{Comparison of methods for identifying altered regulatory regions}
\usage{
comparePeaksAltre(analysisresults, reference, lfctypespecific = 1.5,
  lfcshared = 1.2, pvaltypespecific = 0.01, pvalshared = 0.05)
}
\arguments{
\item{analysisresults}{analysisresults of countanalysis.}

\item{reference}{cell type to be considered "reference" or "reference" to which
other cell types will be compared}

\item{lfctypespecific}{log2fold change for type specific TSS-dists/TSS-proxs}

\item{lfcshared}{log2fold chance for shared TSS-dists/TSS-proxs}

\item{pvaltypespecific}{p-value for type specific TSS-dists/TSS-proxs}

\item{pvalshared}{p-value for shared TSS-dists/TSS-proxs}
}
\value{
matrix comparing the two methods of identifying altered regulatory
regions, one based on peak intensity, the other on peak presence as
determined by hotspot calling algorithms.
}
\description{
Creates a table to compare two methods of identifying altered regulatory
regions, one based on peak intensity, the other on peak presence as
 determined by hotspot calling algorithms.
}
\examples{
\dontrun{
csvfile <- file.path(dir="yourfilepath", 'sampleinfo.csv')
sampleinfo <- loadCSVFile(csvfile)
samplePeaks <- loadBedFiles(sampleinfo)
consPeaks <- getConsensusPeaks(samplepeaks = samplePeaks, minreps = 2)
plotConsensusPeaks(samplepeaks = consPeaks)
TSSannot <- getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consPeaks,
                                           TSS = TSSannot,
                                           merge = TRUE,
                                           regionspecific = TRUE,
                                           distancefromTSSdist = 1500,
                                           distancefromTSSprox = 1000 )
counts_consPeaks <- getCounts(annotpeaks = consPeaksAnnotated,
			sampleinfo=sampleinfo,
                              reference = 'SAEC')
altre_peaks <- countanalysis(counts = counts_consPeaks,
                             pval = 0.01,
                             lfcvalue = 1)
categaltre_peaks <- categAltrePeaks(altre_peaks,
                                    lfctypespecific = 1.5,
                                    lfcshared = 1.2,
                                    pvaltypespecific = 0.01,
                                    pvalshared = 0.05)
analysisresults <- comparePeaksAltre(categaltre_peaks, reference= 'SAEC')
}
}

