% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{enrichHeatmap}
\alias{enrichHeatmap}
\title{Given the output from enrichment(), creates a heatmap from
the ouput of the enrichment analysis. Presence or
absence of the pathway in enrichment of both type-specific (increased
or decreased log2fold change, low p-value) and shared (no change, higher p-value)
regulatory regions is plotted.}
\usage{
enrichHeatmap(input, title, pvalfilt = 0.001, removeonlyshared = FALSE)
}
\arguments{
\item{input}{results from enrichment analysis}

\item{title}{title of the heatmap}

\item{pvalfilt}{p-value cut-off for inclusion in heatmap}

\item{removeonlyshared}{removes regions that come up signifigant only shared
regulatory regions when set to TRUE. Default is FALSE.}
}
\value{
heatmap
}
\description{
Given the output from enrichment(), creates a heatmap from
the ouput of the enrichment analysis. Presence or
absence of the pathway in enrichment of both type-specific (increased
or decreased log2fold change, low p-value) and shared (no change, higher p-value)
regulatory regions is plotted.
}
\examples{
\dontrun{
dir <- system.file("extdata", package="ALTRE", mustWork=TRUE)
csvfile <- file.path(dir, "lung.csv")
samplePeaks <- loadPeaks(csvfile)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
plotConsensusPeaks(samplepeaks=consPeaks)
TSSannot=getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks=consPeaks, TSS=TSSannot,merge=TRUE,
regionspecific=TRUE,mergedistenh=1500,mergedistprom=1000 )
counts_consPeaks <- getcounts(annotpeaks=consPeaksAnnotated, csvfile=csvfile,
reference="SAEC")
 altre_peaks=countanalysis(counts=counts_consPeaks, pval=0.01, lfcvalue=1)
MFenrich=pathenrich(analysisresults=altre_peaks, ontoltype="MF", enrichpvalfilt=0.01)
BPenrich=pathenrich(analysisresults=altre_peaks, ontoltype="BP", enrichpvalfilt=0.01)
plot1=enrichHeatmap(MFenrich, title="GO:MF, p<0.01")
plot2=enrichHeatmap(BPenrich, title="GO:BP, p<0.01")
}
}

