% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcounts.R, R/getcounts2.R
\name{getcounts}
\alias{getcounts}
\title{Generates count data for regulatory regions}
\usage{
getcounts(annotpeaks, csvfile, reference, chrom = NULL)

getcounts(annotpeaks, csvfile, reference, chrom = NULL)
}
\arguments{
\item{annotpeaks}{list output from combineannotatepeaks function}

\item{csvfile}{name of CSV file, with complete path, that contains the following 4 columns for each sample 1) complete filepath; 2) name of bamfiles; 3) name of peak files; 4) name or type of sample; 5) name of replicate}

\item{reference}{name of sample type to be considered "reference" in DESeq2 analysis}

\item{chrom}{optional, only chromosome chrom will be evaluated}

\item{annotpeaks}{list output from combineannotatepeaks function}

\item{csvfile}{name of CSV file, with complete path, that contains the following 4 columns for each sample 1) complete filepath; 2) name of bamfiles; 3) name of peak files; 4) name or type of sample; 5) name of replicate}

\item{reference}{name of sample type to be considered "reference" in DESeq2 analysis}

\item{chrom}{optional, only chromosome chrom will be evaluated}
}
\value{
List containing three items:
(1) DESeqDataSet: contains count information for all replicates of all samples
(2) Matrix: contains number of enhancers and promoters before and after filtering (if applicable)
(3) Plot: contains density plot

List containing three items:
(1) DESeqDataSet: contains count information for all replicates of all samples
(2) Matrix: contains number of enhancers and promoters before and after filtering (if applicable)
(3) Plot: contains density plot
}
\description{
Counts the number of reads in each regulatory region for each sample type --
read count is derived from user-input BAM filex, and regions of interest
are supplied in a Granges object, ideally output of combineannotatepeaks.R.

Counts the number of reads in each regulatory region for each sample type --
read count is derived from user-input BAM filex, and regions of interest
are supplied in a Granges object, ideally output of combineannotatepeaks.R.
}
\examples{
TSSannot <- getTSS()
dir <- system.file("extdata", package="ALTRE", mustWork=TRUE)
csvfile <- file.path(dir, "lung.csv")
samplePeaks <- loadPeaks(csvfile)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
consPeaksAnnotated=combineAnnotatePeaks(conspeaks=consPeaks, TSS=TSSannot)
counts_consPeaks=getcounts(annotpeaks=consPeaksAnnotated, csvfile=csvfile, reference="SAEC", chrom="chr21")
dir=system.file("extdata", package="ALTRE", mustWork=TRUE)
TSSpath=file.path(dir,"Homosapiens_GRCh37.75_TSS.bed")
TSSannot=read.table(TSSpath, header=TRUE)
dir <- system.file("extdata", package="ALTRE", mustWork=TRUE)
csvfile <- file.path(dir, "lung.csv")
samplePeaks <- loadPeaks(csvfile)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
consPeaksAnnotated=combineAnnotatePeaks(conspeaks=consPeaks, TSS=TSSannot)
counts_consPeaks=getcounts2(annotpeaks=consPeaksAnnotated, csvfile=csvfile, reference="SAEC", chrom=NULL)
}

