% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_getcounts.R
\name{getCounts}
\alias{getCounts}
\title{Generates count data for regulatory regions}
\usage{
getCounts(annotpeaks, sampleinfo, reference, singleEnd = TRUE, chrom = NULL)
}
\arguments{
\item{annotpeaks}{list output from combineannotatepeaks function}

\item{sampleinfo}{dataframe as returned from loadCSVFile() function}

\item{reference}{name of sample type to be
considered 'reference' in DESeq2 analysis}

\item{singleEnd}{whether input data is single-end (default is TRUE)}

\item{chrom}{optional, only chromosome chrom will be evaluated}
}
\value{
List containing three items:
(1) DESeqDataSet: contains count information for all replicates of all samples
(2) Matrix: contains number of TSS-distal and TSS-proximal
 before and after filtering (if applicable)
(3) Data frame for creating a density plot (use function plotgetcounts()
}
\description{
Counts the number of reads in each regulatory region for each sample type --
read count is derived from user-input BAM filex, and regions of interest
are supplied in a Granges object, ideally output of combineannotatepeaks.R.
}
\examples{
\dontrun{
csvfile <- loadCSVFile("DNAseEncodeExample.csv")
samplePeaks <- loadBedFiles(csvfile)
consensusPeaks <- getConsensusPeaks(samplepeaks = samplePeaks, minreps = 2)
TSSannot <- getTSS()
consensusPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consensusPeaks,
   TSS = TSSannot,
   merge = TRUE,
   regionspecific = TRUE,
   distancefromTSSdist = 1500,
   distancefromTSSprox = 1000)
consensusPeaksCounts <- getCounts(annotpeaks = consensusPeaksAnnotated,
   sampleinfo = csvfile,
   reference = 'SAEC',
   chrom = 'chr21')
}
}

