% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{list of plots}

\item{plotlist}{plotlist}

\item{file}{file}

\item{cols}{number of columns in layout}

\item{layout}{matrix specifying layout, if present cols is ignored}
}
\description{
Plots multiple ggplot objects in one window.
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
This function was not written by the authours of this package. Link:
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\examples{
\dontrun{
dir <- system.file('extdata', package='ALTRE', mustWork=TRUE)
csvfile <- file.path(dir, 'lung.csv')
sampleinfo <- loadCSVFile(csvfile)
samplePeaks <- loadBedFiles(sampleinfo)
consPeaks <- getConsensusPeaks(samplepeaks = samplePeaks, minreps = 2)
plotConsensusPeaks(samplepeaks = consPeaks)
TSSannot <- getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consPeaks,
                                          TSS = TSSannot,
                                          merge = TRUE,
                                          regionspecific = TRUE,
                                          distancefromTSSdist = 1500,
                                          distancefromTSSprox = 1000 )
counts_consPeaks <- getCounts(annotpeaks = consPeaksAnnotated,
                             sampleinfo = sampleinfo,
                             reference = 'SAEC',
                             chrom = 'chr21')
altre_peaks <- countanalysis(counts = counts_consPeaks,
                             pval = 0.01,
                             lfcvalue = 1)
categaltre_peaks <- categAltrePeaks(altre_peaks,
                                    lfctypespecific = 1.5,
                                    lfcshared = 1.2,
                                    pvaltypespecific = 0.01,
                                    pvalshared = 0.05)
MFenrich <- pathenrich(analysisresults = categaltre_peaks,
                      ontoltype = 'MF',
                      enrichpvalfilt = 0.01)
BPenrich <- pathenrich(analysisresults= categaltre_peaks,
                      ontoltype='BP',
                      enrichpvalfilt=0.01)
plot1 <- enrichHeatmap(MFenrich, title='GO:MF, p<0.01')
plot2 <- enrichHeatmap(BPenrich, title='GO:BP, p<0.01')
multiplot(plot1,plot2,cols=1)
}

}

