% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categaltrepeaks.R, R/enrichment.R
\name{pathenrich}
\alias{pathenrich}
\title{Categorize altered peaks as experiment-specific, reference-specific, or shared}
\usage{
pathenrich(analysisresults, ontoltype = "MF", enrichpvalfilt = 0.01,
  lfctypespecific = 1.5, lfcshared = 1.2, pvaltypespecific = 0.01,
  pvalshared = 0.05, genes = 20, offspring = 300,
  regionsubset = "promoter")

pathenrich(analysisresults, ontoltype = "MF", enrichpvalfilt = 0.01,
  lfctypespecific = 1.5, lfcshared = 1.2, pvaltypespecific = 0.01,
  pvalshared = 0.05, genes = 20, offspring = 300,
  regionsubset = "promoter")
}
\arguments{
\item{analysisresults}{results from analysis of countanalysis()}

\item{ontoltype}{one of three categories: "MF" (molecular function), "CC" (cellular component), "BP" (biological process)}

\item{enrichpvalfilt}{adjusted pval for enrichment to filter on (adjusted for multiple testing)}

\item{lfctypespecific}{log2fold change (of chromatin accessibility) for type specific enhancers/promoters}

\item{lfcshared}{log2fold change (of chromatin accessibility) for shared enhancers/promoters}

\item{pvaltypespecific}{p-value (of chromatin accessibility) for type specific enhancers/promoters}

\item{pvalshared}{p-value (of chromatin accessibility) for shared enhancers/promoters}

\item{genes}{minimum number of genes allowable in a pathway}

\item{offspring}{maximum number of offspring allowable in a pathway}

\item{regionsubset}{"promoter" or "enhancer"}

\item{analysisresults}{results from analysis of counts}

\item{lfctypespecific}{log2fold change (of chromatin accessibility) for type specific enhancers/promoters}

\item{lfcshared}{log2fold change (of chromatin accessibility) for shared enhancers/promoters}

\item{pvaltypespecific}{p-value (of chromatin accessibility) for type specific enhancers/promoters}

\item{pvalshared}{p-value (of chromatin accessibility) for shared enhancers/promoters}
}
\value{
list 1) stats (number of experiment-specific, reference-specific, and shared REs); 
	2) a volcano plot

dataframe identifying p-values for enriched pathways -- pathways also annotated with additional information
}
\description{
Categorize altered peaks as experiment-specific, reference-specific, or shared

Determine which pathways are overrepresented in the
altered promoters and enhancers. Pathways are determined by linking
the enhancer/promoter to the nearest gene, and then linking genes to pathways using
the gene ontology database. The "gene" argument limits how few genes a pathway can contain,
while the "offspring" argument limits how many offspring a pathway can contain.
Pathways with low gene counts are less reliable (often false positives), while pathways
with many offspring are vague and unlikely to be of much use -- the enrichment of
their more precise offspring is the more interesting question.
************CHANGE ANALYSISRESULTS TO ALTREPEAKS!!!!******************
****************CHANGE REGIONSUBSET TO REGIONTYPE!!!!*****************
}
\examples{
dir <- system.file("extdata", package="ALTRE", mustWork=TRUE)
csvfile <- file.path(dir, "lung.csv")
samplePeaks <- loadPeaks(csvfile)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
TSSannot <- getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks=consPeaks, TSS=TSSannot,merge=TRUE,
regionspecific=TRUE,mergedistenh=1500,mergedistprom=1000 )
# Need to run getcounts on all chromosomes
counts_consPeaks=getcounts(annotpeaks=consPeaksAnnotated, csvfile=csvfile, reference="SAEC")
altre_peaks=countanalysis(counts=counts_consPeaks, pval=0.01, lfcvalue=1)
categaltre_peaks=categAltrePeaks(altrepeaks, lfctypespecific=1.5,lfcshared=1.2,
pvaltypespecific=0.01,pvalshared=0.05)

dir <- system.file("extdata", package="ALTRE", mustWork=TRUE)
csvfile <- file.path(dir, "lung.csv")
samplePeaks <- loadPeaks(csvfile)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
TSSannot <- getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks=consPeaks, TSS=TSSannot,merge=TRUE,
regionspecific=TRUE,mergedistenh=1500,mergedistprom=1000 )
# Need to run getcounts on all chromosomes
counts_consPeaks=getcounts(annotpeaks=consPeaksAnnotated, csvfile=csvfile, reference="SAEC")
altre_peaks=countanalysis(counts=counts_consPeaks, pval=0.01, lfcvalue=1)
MFenrich=pathenrich(analysisresults=altre_peaks, ontoltype="MF", enrichpvalfilt=0.01)
BPenrich=pathenrich(analysisresults=altre_peaks, ontoltype="BP", enrichpvalfilt=0.01)

}

