% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotCombineAnnotatePeaks}
\alias{plotCombineAnnotatePeaks}
\title{Given the output from combineAnnotatePeaks,
plot a barplot showing number of peaks before/after merging
(only works if peaks were merged)}
\usage{
plotCombineAnnotatePeaks(conspeaks)
}
\arguments{
\item{conspeaks}{output generated from combineAnnotatePeaks}
}
\value{
a ggplot
}
\description{
Given the output from combineAnnotatePeaks,
plot a barplot showing number of peaks before/after merging
(only works if peaks were merged)
}
\examples{
\dontrun{
dir <- system.file('extdata', package='ALTRE', mustWork=TRUE)
csvfile <- file.path(dir, 'lung.csv')
samplePeaks <- loadPeaks(csvfile)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
plotConsensusPeaks(samplepeaks=consPeaks)
TSSannot=getTSS()
 consPeaksAnnotated <- combineAnnotatePeaks(conspeaks=consPeaks,
 TSS=TSSannot, merge=TRUE, regionspecific=TRUE, mergedistenh=1500,
  mergedistprom=1000)
plotCombineAnnotatePeaks(consPeaksAnnotated)
}
}

