% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotCombineAnnotatePeaks}
\alias{plotCombineAnnotatePeaks}
\title{Given the output from combineAnnotatePeaks,
plot a barplot showing number of peaks before/after merging
(only works if peaks were merged)}
\usage{
plotCombineAnnotatePeaks(conspeaks)
}
\arguments{
\item{conspeaks}{output generated from combineAnnotatePeaks}
}
\value{
a ggplot
}
\description{
Given the output from combineAnnotatePeaks,
plot a barplot showing number of peaks before/after merging
(only works if peaks were merged)
}
\examples{
\dontrun{
dir <- system.file('extdata', package='ALTRE', mustWork=TRUE)
csvfile <- file.path(dir, 'lung.csv')
sampleinfo <- loadCSVFile(csvfile)
samplePeaks <- loadBedFiles(sampleinfo)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
plotConsensusPeaks(samplepeaks=consPeaks)
TSSannot <- getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consPeaks,
                                          TSS = TSSannot,
                                          merge = TRUE,
                                          regionspecific = TRUE,
                                          distancefromTSSdist = 1500,
                                          distancefromTSSprox = 1000)
plotCombineAnnotatePeaks(consPeaksAnnotated)
}
}

