% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotCompareMethods}
\alias{plotCompareMethods}
\title{Plots a venn diagram that compares altered regions as determined by peak presence or by #' differential counts.  The type of regulatory region (TSS-proximal, TSS-distal, or both)
and type of peak comparison (intensity or peak) must be specified.
Plots a venn diagram that compares altered regions as determined by peak
presence or by differential counts.  The type of regulatory region
(TSS-proximal, TSS-distal, or both) and type of peak comparison
(intensity or peak) must be specified.}
\usage{
plotCompareMethods(analysisresultsmatrix, region = "both",
  method = "Intensity", palette = NULL, maintitle = "default",
  maintitlesize = "20px")
}
\arguments{
\item{analysisresultsmatrix}{analysisresults of Intensity analysis place into
analysisresults matrix by the analyzeanalysisresults function}

\item{region}{pick a region, regions can be 'TSS-distal', 'TSS-proximal',
or 'both' -- INCLUDE quotes}

\item{method}{pick a method, methods can be 'Intensity' or 'Peak'
include quotes}

\item{palette}{RColorBrewer palette to change graph colors}

\item{maintitle}{main title (default generated from sample names)}

\item{maintitlesize}{main title size (default, 20px)}
}
\value{
venn diagram
}
\description{
Plots a venn diagram that compares altered regions as determined by peak presence or by #' differential counts.  The type of regulatory region (TSS-proximal, TSS-distal, or both)
and type of peak comparison (intensity or peak) must be specified.
Plots a venn diagram that compares altered regions as determined by peak
presence or by differential counts.  The type of regulatory region
(TSS-proximal, TSS-distal, or both) and type of peak comparison
(intensity or peak) must be specified.
}
\examples{
\dontrun{
csvfile <- loadCSVFile("DNAseEncodeExample.csv")
samplePeaks <- loadBedFiles(csvfile)
consensusPeaks <- getConsensusPeaks(samplepeaks = samplePeaks,
minreps = 2)
TSSannot <- getTSS()
consensusPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consensusPeaks,
TSS = TSSannot,
merge = TRUE,
regionspecific = TRUE,
distancefromTSSdist = 1500,
distancefromTSSprox = 1000)
consensusPeaksCounts <- getCounts(annotpeaks = consensusPeaksAnnotated,
                                 sampleinfo = csvfile,
                                 reference = 'SAEC',
                                 chrom = 'chr21')
alteredPeaks <- countanalysis(counts = consensusPeaksCounts,
pval = 0.01,
lfcvalue = 1)
alteredPeaksCategorized <- categAltrePeaks(alteredPeaks,
                                          lfctypespecific = 1.5,
                                          lfcshared = 1.2,
                                          pvaltypespecific = 0.01,
                                          pvalshared = 0.05)
comparePeaksAnalysisResults <- comparePeaksAltre(alteredPeaksCategorized)
plotCompareMethods(comparePeaksAnalysisResults)
}
}

