% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotCompareMethodsAll}
\alias{plotCompareMethodsAll}
\title{Plots pie charts for comparison of two methods of identifying altered
regulatory regions.  Makes pie charts for TSS-proximal, TSS-distal, and
combined for both intensity-based peaks and for peaks identified by hotspot
calling algorithms.  There is no return value. Six pie charts swill be
plotted.}
\usage{
plotCompareMethodsAll(analysisresultsmatrix, viewer = TRUE,
  palette = c("#C71585", "#d3d3d3", "#00E5EE", "#000080"), title11 = NULL,
  title12 = NULL, title13 = NULL, title21 = NULL, title22 = NULL,
  title23 = NULL, maintitlesize = "20px")
}
\arguments{
\item{analysisresultsmatrix}{analysisresults of countanalysis function
place into a a analysisresults matrix by the analyzeanalysisresults function}

\item{viewer}{whether the plot should be displayed in the RStudio viewer or
in Shiny/Knittr}

\item{palette}{choose an RColorBrewer palette ("Set1", "Set2", "Set3", "Pastel1",
"Pastel2", "Paired", etc.) or submit a vector of colors}

\item{title11}{title of the first graph in the first row}

\item{title12}{title of the second graph in the first row}

\item{title13}{title of the third graph in the first row}

\item{title21}{title of the first graph in the second row}

\item{title22}{title of the second graph in the second row}

\item{title23}{title of the third graph in the second row}

\item{maintitlesize}{main title size (default, 20px)}
}
\description{
Plots pie charts for comparison of two methods of identifying altered
regulatory regions.  Makes pie charts for TSS-proximal, TSS-distal, and
combined for both intensity-based peaks and for peaks identified by hotspot
calling algorithms.  There is no return value. Six pie charts swill be
plotted.
}
\examples{
\dontrun{
csvfile <- loadCSVFile("DNaseEncodeExample.csv")
samplePeaks <- loadBedFiles(csvfile)
consensusPeaks <- getConsensusPeaks(samplepeaks = samplePeaks,
minreps = 2)
TSSannot <- getTSS()
consensusPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consensusPeaks,
TSS = TSSannot,
merge = TRUE,
regionspecific = TRUE,
distancefromTSSdist = 1500,
distancefromTSSprox = 1000)
consensusPeaksCounts <- getCounts(annotpeaks = consensusPeaksAnnotated,
                                 sampleinfo = csvfile,
                                 reference = 'SAEC',
                                 chrom = 'chr21')
alteredPeaks <- countanalysis(counts = consensusPeaksCounts,
pval = 0.01,
lfcvalue = 1)
alteredPeaksCategorized <- categAltrePeaks(alteredPeaks,
                                          lfctypespecific = 1.5,
                                          lfcshared = 1.2,
                                          pvaltypespecific = 0.01,
                                          pvalshared = 0.05)
comparePeaksAnalysisResults <- comparePeaksAltre(alteredPeaksCategorized)
plotCompareMethodsAll(comparePeaksAnalysisResults)
}
}
