% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgetconsensuspeaks.R
\name{plotConsensusPeaks}
\alias{plotConsensusPeaks}
\title{Given the output from getConsensusPeaks, generate a barplot of count statistics}
\usage{
plotConsensusPeaks(samplepeaks)
}
\arguments{
\item{samplepeaks}{output generated from getConsensusPeaks}
}
\value{
a ggplot
}
\description{
Given the output from getConsensusPeaks, generate a barplot of count statistics
}
\examples{
dir <- system.file("extdata", package="ALTRE", mustWork=TRUE)
csvfile <- file.path(dir, "lung.csv")
samplePeaks <- loadPeaks(csvfile)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
plotConsensusPeaks(samplepeaks=consPeaks)

}

