% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotConsensusPeaks}
\alias{plotConsensusPeaks}
\title{Given the output from getConsensusPeaks, generate a barplot
of countstatistics}
\usage{
plotConsensusPeaks(samplepeaks, viewer = TRUE, palette = "Set1",
  xlabel = NULL, ylabel = "Peak Counts", xlabelsize = "15px",
  ylabelsize = "15px", maintitle = "Peak Counts by Cell Type",
  subtitle = "For bioreplicates and their merged consensus track",
  maintitlesize = "20px", subtitlesize = "15px")
}
\arguments{
\item{samplepeaks}{output generated from getConsensusPeaks}

\item{viewer}{whether the plot should be displayed in the RStudio viewer or
in Shiny/Knittr}

\item{palette}{choose an RColorBrewer palette ("Set1", "Set2", "Set3",
"Pastel1", "Pastel2", "Paired", etc.) or submit a vector of colors}

\item{xlabel}{label for x-axis (default, types of peaks - e.g.
ConsensusPeaks, rep1, rep2, etc.)}

\item{ylabel}{label for y-axis (default, "Peak Counts")}

\item{xlabelsize}{size of xlabel (default, 15px)}

\item{ylabelsize}{size of ylabel (default, 15px)}

\item{maintitle}{main title (default, "Peak Counts by Cell Type")}

\item{subtitle}{subtitle (default, "For bioreplicates and their merged
consensus track")}

\item{maintitlesize}{main title size (default, 20px)}

\item{subtitlesize}{subitle size (default 15px)}
}
\value{
a highcharter object
}
\description{
Given the output from getConsensusPeaks, generate a barplot
of countstatistics
}
\examples{
\dontrun{
csvfile <- loadCSVFile("DNAseEncodeExample.csv")
samplePeaks <- loadBedFiles(csvfile)
consensusPeaks <- getConsensusPeaks(samplepeaks = samplePeaks,
minreps = 2)
plotConsensusPeaks(samplepeaks = consensusPeaks)}
}

