% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotConsensusPeaks}
\alias{plotConsensusPeaks}
\title{Given the output from getConsensusPeaks, generate a barplot
of countstatistics}
\usage{
plotConsensusPeaks(samplepeaks, cols = c("#0099ff", "#ff5050"))
}
\arguments{
\item{samplepeaks}{output generated from getConsensusPeaks}

\item{cols}{hex colors for points in this order: Experiment-Specific, Reference-Specific, Shared}
}
\value{
a highcharter object
}
\description{
Given the output from getConsensusPeaks, generate a barplot
of countstatistics
}
\examples{
\dontrun{
csvfile <- file.path(dir="yourfilepath", 'sampleinfo.csv')
sampleinfo <- loadCSVFile(csvfile)
samplePeaks <- loadBedFiles(sampleinfo)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks, minreps=2)
plotConsensusPeaks(samplepeaks = consPeaks)
}
}

