% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotCountAnalysis}
\alias{plotCountAnalysis}
\title{Create a volcano plot from the output of categAltrePeaks}
\usage{
plotCountAnalysis(altrepeakscateg, viewer = TRUE, palette = c("#C71585",
  "#d3d3d3", "#00E5EE", "#000080"), maintitlelefty = "TSS-distal",
  maintitlerighty = "TSS-proximal", ylabel = "-log10 pvalue",
  xlabel = "log2fold change", xlabelsize = "15px", ylabelsize = "15px",
  maintitlesize = "20px")
}
\arguments{
\item{altrepeakscateg}{output generated from countanalysis() then
categAltrePeaks()}

\item{viewer}{whether the plot should be displayed in the RStudio viewer or
in Shiny/Knittr}

\item{palette}{choose an RColorBrewer palette ("Set1", "Set2", "Set3",
"Pastel1", "Pastel2", "Paired", etc.) or submit a vector of colors}

\item{maintitlelefty}{main title (default, "TSS-distal")}

\item{maintitlerighty}{main title (default, "TSS-proximal")}

\item{ylabel}{label for y-axis (default, "Density")}

\item{xlabel}{label for x-axis (default, "-log10 pvalue")}

\item{xlabelsize}{size of xlabel (default, 15px)}

\item{ylabelsize}{size of ylabel (default, 15px)}

\item{maintitlesize}{main title size (default, 20px)}
}
\value{
a highcharter object
}
\description{
Create a volcano plot from the output of categAltrePeaks
}
\examples{
\dontrun{
csvfile <- loadCSVFile("DNAseEncodeExample.csv")
samplePeaks <- loadBedFiles(csvfile)
consensusPeaks <- getConsensusPeaks(samplepeaks = samplePeaks, minreps = 2)
TSSannot <- getTSS()
consensusPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consensusPeaks,
                                          TSS = TSSannot,
                                          merge = TRUE,
                                          regionspecific = TRUE,
                                          distancefromTSSdist = 1500,
                                          distancefromTSSprox = 1000)
consensusPeaksCounts <- getCounts(annotpeaks = consensusPeaksAnnotated,
                              reference = 'SAEC',
                              sampleinfo = csvfile,
                              chrom = 'chr21')
alteredPeaks <- countanalysis(counts=consensusPeaksCounts,
                             pval=0.01,
                             lfcvalue=1)
alteredPeaksCategorized <- categAltrePeaks(alteredPeaks,
                             lfctypespecific = 1.5,
                             lfcshared = 1.2,
                             pvaltypespecific = 0.01,
                             pvalshared = 0.05)
plotCountAnalysis(altrepeakscateg = alteredPeaksCategorized)
}
}

