% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcountanalysis.R
\name{plotCountAnalysis}
\alias{plotCountAnalysis}
\title{Given the output from getcounts, plot a density plot of log2 RPKM values of regulation regions}
\usage{
plotCountAnalysis(altrepeaks)
}
\arguments{
\item{countsconspeaks}{output generated from getcounts}
}
\value{
a ggplot
}
\description{
Given the output from getcounts, plot a density plot of log2 RPKM values of regulation regions
}
\examples{
dir <- system.file("extdata", package="ALTRE", mustWork=TRUE)
csvfile <- file.path(dir, "lung.csv")
samplePeaks <- loadPeaks(csvfile)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
plotConsensusPeaks(samplepeaks=consPeaks)
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks=consPeaks, TSS=TSSannot)
counts_consPeaks <-getcounts(annotpeaks=consPeaksAnnotated, csvfile=csvfile,
reference="SAEC", chrom="chr21")
altre_peaks=countanalysis(counts=counts_consPeaks, pval=0.01, lfcvalue=1)
plotCountAnalysis(altre_peaks)

}

