% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotDistCountAnalysis}
\alias{plotDistCountAnalysis}
\title{Creates a boxplot to see the distribution of read counts in type-specific and shared enhancers}
\usage{
plotDistCountAnalysis(analysisresults, counts)
}
\arguments{
\item{analysisresults}{output generated from countanalysis() then categAltrePeaks()}

\item{counts}{output generated from getcounts()}
}
\value{
a ggplot
}
\description{
Takes the rlog transformation of the RRKM (Reads Per Kilobase of transcript per Million)
of the read counts of type-specific and shared regulatory regions and plots
the distribution of those read counts in all sample types analyzed in the workflow.
}
\examples{
\dontrun{
dir <- system.file('extdata', package='ALTRE', mustWork=TRUE)
csvfile <- file.path(dir, 'lung.csv')
sampleinfo <- loadCSVFile(csvfile)
samplePeaks <- loadBedFiles(sampleinfo)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
plotConsensusPeaks(samplepeaks=consPeaks)
TSSannot<- getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consPeaks,
                                          TSS = TSSannot,
                                          merge = TRUE,
                                          regionspecific = TRUE,
                                          mergedistenh = 1500,
                                          mergedistprom = 1000)
counts_consPeaks <- getcounts(annotpeaks = consPeaksAnnotated,
                              sampleinfo = sampleinfo,
                              reference = 'SAEC',
                              chrom = 'chr21')
altre_peaks <- countanalysis(counts = counts_consPeaks,
                             pval = 0.01,
                             lfcvalue = 1)
categaltre_peaks <- categAltrePeaks(altre_peaks,
                                    lfctypespecific = 1.5,
                                    lfcshared = 1.2,
                                    pvaltypespecific = 0.01,
                                    pvalshared = 0.05)
plotDistCountAnalysis(categaltre_peaks, counts_consPeaks)
}
}

