% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotGetCounts}
\alias{plotGetCounts}
\title{Given the output from getCounts, plot a density plot of
log2 RPKM values of regulation regions}
\usage{
plotGetCounts(countsConsPeaks)
}
\arguments{
\item{countsConsPeaks}{output generated from getCounts}
}
\value{
a ggplot
}
\description{
Given the output from getCounts, plot a density plot of
log2 RPKM values of regulation regions
}
\examples{
\dontrun{
csvfile <- file.path(dir="yourfilepath", 'sampleinfo.csv')
sampleinfo <- loadCSVFile(csvfile)
samplePeaks <- loadBedFiles(sampleinfo)
consPeaks <- getConsensusPeaks(samplepeaks = samplePeaks, minreps=2)
plotConsensusPeaks(samplepeaks = consPeaks)
TSSannot <- getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consPeaks,
                                          TSS = TSSannot,
                                          merge = TRUE,
                                          regionspecific = TRUE,
                                          distancefromTSSdist = 1500,
                                          distancefromTSSprox = 1000 )

counts_consPeaks <- getCounts(annotpeaks = consPeaksAnnotated,
                             sampleinfo = sampleinfo,
                             reference = 'SAEC',
                             chrom = 'chr21')
plotgetcounts(counts_consPeaks)
}
}

