% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{plotGetCounts}
\alias{plotGetCounts}
\title{Given the output from getCounts, plot a density plot of
log2 RPKM values of regulation regions}
\usage{
plotGetCounts(countsConsPeaks, palette = "Set1")
}
\arguments{
\item{countsConsPeaks}{output generated from getCounts}

\item{palette}{RColorBrewer palette to change graph colors}
}
\value{
a highcharter object
}
\description{
Given the output from getCounts, plot a density plot of
log2 RPKM values of regulation regions
}
\examples{
\dontrun{
csvfile <- loadCSVFile("DNAseEncodeExample.csv")
samplePeaks <- loadBedFiles(csvfile)
consensusPeaks <- getConsensusPeaks(samplepeaks = samplePeaks,
minreps = 2)
TSSannot <- getTSS()
consensusPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consensusPeaks,
TSS = TSSannot,
merge = TRUE,
regionspecific = TRUE,
distancefromTSSdist = 1500,
distancefromTSSprox = 1000)
consensusPeaksCounts <- getCounts(annotpeaks = consensusPeaksAnnotated,
                                 sampleinfo = csvfile,
                                 reference = 'SAEC',
                                 chrom = 'chr21')
plotGetCounts(consensusPeaksCounts)}
}

