% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8a_runGREAT.R
\name{processPathways}
\alias{processPathways}
\title{Enrichment analysis using GREAT package
to identify putative pathways of interest for further
investigation}
\usage{
processPathways(GREATpath, pathway_category = "GO", adjustby = "bonferroni",
  test = "Binom", enrichcutoff = 2, adjpvalcutoff = 0.05)
}
\arguments{
\item{GREATpath}{output of runGREAT()}

\item{pathway_category}{character, "GO", "Pathway Data", "Regulatory Motifs",
"Phenotype Data and Human Disease", "Gene Expression", "Gene Families"
(default is "GO")}

\item{adjustby}{character, "fdr" or "bonferroni", default is "bonferroni"}

\item{test}{character, "Both" denotes hypergeometric and binomical tests are used to
    determine enriched pathways, "Binom" denotes binomial tests used, "Hyper" denotes
hypergeometric tests are used.  Default is "Binom"}

\item{enrichcutoff}{numeric, fold change enrichment cutoff to determine enriched pathways,
default is 2}

\item{adjpvalcutoff}{numeric, Bonferroni adjusted p-value cutoff to determine enriched pathways,
default is 0.05}
}
\value{
list of dataframes for enriched pathways - each dataframe in the list
represents one pathway type (e.g. "GO Molecular Function")
}
\description{
Enrichment analysis using GREAT package
to identify putative pathways of interest for further
investigation
}

