% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalfunctions.R
\name{tssannotgrange}
\alias{tssannotgrange}
\title{Internal function: annotates peaks as TSS-proximals or TSS-distals}
\usage{
tssannotgrange(grange, TSS, distancefromTSS)
}
\arguments{
\item{grange}{GRanges object}

\item{TSS}{GRanges of Transcription Start Site Locations}

\item{distancefromTSS}{distancefromTSS}
}
\value{
GRanges object
}
\description{
Takes a Grange input
Not for use by package user
}

