% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writerfunctions.R
\name{writeAnnotatedRegions}
\alias{writeAnnotatedRegions}
\title{Writes annotated genomic regions}
\usage{
writeAnnotatedRegions(countsPeaks, con)
}
\arguments{
\item{countsPeaks}{output generated from getCounts()}

\item{con}{connection filepath}
}
\description{
Creates a CSV file with genomic regions annoted as promotor or enhances
categorized by cell type
}
\examples{
\dontrun{
dir <- system.file('extdata', package='ALTRE', mustWork=TRUE)
csvfile <- file.path(dir, 'lung.csv')
sampleinfo <- loadCSVFile(csvfile)
samplePeaks <- loadBedFiles(sampleinfo)
consPeaks <- getConsensusPeaks(samplepeaks=samplePeaks,minreps=2)
TSSannot <- getTSS()
consPeaksAnnotated <- combineAnnotatePeaks(conspeaks = consPeaks,
                                          TSS = TSSannot,
                                          merge = TRUE,
                                          regionspecific = TRUE,
                                          mergedistenh = 1500,
                                          mergedistprom = 1000 )
#Need to run getcounts on all chromosomes
countsPeaks <- getcounts(annotpeaks = consPeaksAnnotated,
                             csvfile = csvfile,
                             reference = 'SAEC')
con <- "annotatedRegions.csv"
writeAnnotatedRegions(countsPeaks, con)
}
}

